fake-firmware-test: async firmware use after free test
This provides a test against the issue reported and
fixed on firmware_class use of schedule_work() for
asynchronous behaviour when taking the name of the
desired firmware from the caller. The caller is allowed
to pass a pointer to the firmware name, this can however
immediately be freed by the caller and when that happens
can cause a use-after-free scenario.

In order to trigger the issue faster you can apply the
supplied patch in the tree force-use-after-free.patch
to your Linux kernel sources. Then just build this
driver and load it.

With a fixed kernel you should see:

[  283.512445] firmware name: test_module_stuff.bin
[  287.514020] firmware name: test_module_stuff.bin
[  287.532489] firmware found

With a buggy kernel you should see a request for arbitrary
files but in the most common case you kernel will crash
due to a page fault on trying to free something you cannot
and/or NULL pointer dereference.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>